simplify sender interface (#17822)
* simplify sender interface

previously the sender had to return
- a PublishResult
- an Error

The problem here was that PublishResult was treated as an error
indicator for the eventMesh Backend, and the error was for other senders
such as the on Nats

The new interface just expects an Error to be returned in case of an
error. Any successful call will result in a 204 response with an empty
message.

This means now the sender has to map its errors or error responses onto
an error object, and any other function in the chain can just treat all
errors equally.

* remove redundant local variable

* bump images

* fix lint issues

* add missing files

* improve readability of tests

* fix linter

* remove redundant metrics

* empty commit

* add more logging in case of errors during send

* add missing log